+++ CHANGELOG.md
@@ -87,7 +87,7 @@
 ## v3.8.0
 
 - added dot cx, cy in the onDataPointClick functions arguments
-- fixed for horizontal label position when there is only one data point and fromZero prop is true
+- fixed for horizontal label position when there is only one data point and yStart prop is true
 
 ## v3.7.0